Now loads the disc from context if present
